Move tsdbblockutil into tsdbutil package
After a quick look I realized that tsdbblockutil.go was only used by
testing files and that there was an existing package for this same
purpose called tsdbutil.

This commit is simply moving the file into the tsdbutil/block package and
also renaming it to block.go following the internal structure of the package
for other utilitary functions like chunks.go.

Note that it can't be placed directly inside tsdbutil because it imports
storage pkg and storage imports tsdbutil leading to a cyclical
dependency. This is probably the reason why it had the previous
structure in the first place.
